#2600
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@5573 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.testing/src/org/amanzi/testing/AbstractAWETest.java
@@ -70,14 +70,15 @@ public abstract class AbstractAWETest {
     }
     
     private static void deleteDirectory(File directory) {
-        for (File subFile : directory.listFiles()) {
-            if (subFile.isDirectory()) {
-                deleteDirectory(subFile);
-            }            
-            else {
-                subFile.delete();
+        if (directory.exists()) {
+            for (File subFile : directory.listFiles()) {
+                if (subFile.isDirectory()) {
+                    deleteDirectory(subFile);
+                } else {
+                    subFile.delete();
+                }
             }
+            directory.delete();
         }
-        directory.delete();
     }
 }
